+++ crates/chia-sdk-driver/src/primitives/option.rs
@@ -2,9 +2,7 @@ mod option_contract;
 mod option_info;
 mod option_launcher;
 mod option_metadata;
-mod option_singleton;
 
 pub use option_contract::*;
 pub use option_info::*;
 pub use option_metadata::*;
-pub use option_singleton::*;
